move cli logic into cli package to prep for seperate core package
